Personal_Finance_Manager/Root_Directory/requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
mysql-connector-python==8.1.0
Werkzeug==2.3.7
scikit-learn==1.3.0
pandas==2.1.1
numpy==1.24.3
joblib==1.3.2